+++ inst/livy/spark-2.3.0/livyutils.scala
@@ -9,9 +9,10 @@ object LivyUtils {
     "SQLUtils" -> SQLUtils,
     "Utils" -> Utils,
     "Repartition" -> Repartition,
+    "ArrowHelper" -> ArrowHelper,
+    "ArrowConverters" -> ArrowConverters,
     "ApplyUtils" -> ApplyUtils,
     "WorkerHelper" -> WorkerHelper,
-    "WorkerUtils" -> WorkerUtils,
     "MLUtils" -> MLUtils,
     "MLUtils2" -> MLUtils2,
     "BucketizerUtils" -> BucketizerUtils