+++ tests/testthat/test-ml-clustering-bisecting-kmeans.R
@@ -248,7 +248,10 @@ test_that("ml_bisecting_kmeans() works properly", {
   expect_equal(bkm$cluster_centers(), cluster_centers)
 })
 
-test_that("ml_bisecting_kmeans() works properly", {
+test_that("ml_bisecting_kmeans() works properly with iris", {
+  # bisecting kmeans not deterministic over partitions
+  skip_on_arrow()
+
   sc <- testthat_spark_connection()
   test_requires_version("2.0.0", "ml_bisecting_kmeans() requires Spark 2.0.0+")
   iris_tbl <- testthat_tbl("iris")