+++ tests/testthat/test-ml-feature-word2vec.R
@@ -45,10 +45,9 @@ test_that("ml_find_synonyms works properly", {
   tokenized_tbl <- testthat_tbl("tokenized")
   model <- ft_word2vec(sc, "words", "result", vector_size= 3, min_count = 0,
                        dataset = tokenized_tbl)
-  expect_equal(
-    ml_find_synonyms(model, "java", 2) %>%
-      pull(word),
-    c("models", "regression")
-  )
+
+  synonyms <- ml_find_synonyms(model, "java", 2) %>% pull(word)
+
+  expect_true("models" %in% synonyms)
 })